get genesis for all subnets on teleporter deploy
